replaced Zengo Wasm /w RustCrypto impl.  Map reduced the sig shares.
